+++ example/src/jvmMain/kotlin/plugins/Security.kt
@@ -1,0 +1,7 @@
+package plugins
+
+import io.ktor.server.application.*
+
+fun Application.configureSecurity() {
+    configureAppCheck()
+}