+++ example/src/jvmTest/kotlin/ApplicationTest.kt
@@ -1,0 +1,15 @@
+import io.ktor.server.testing.*
+import kotlin.test.Test
+
+class ApplicationTest {
+    @Test
+    fun testRoot() = testApplication {
+//        application {
+//            configureRouting()
+//        }
+//        client.get("/").apply {
+//            assertEquals(HttpStatusCode.OK, status)
+//            assertEquals("Hello World!", bodyAsText())
+//        }
+    }
+}