+++ example/src/main/kotlin/net/freshplatform/plugins/Security.kt
@@ -1,7 +1,0 @@
-package net.freshplatform.plugins
-
-import io.ktor.server.application.*
-
-fun Application.configureSecurity() {
-    configureAppCheck()
-}