+++ scripts/before-push.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+echo "Build with Gradle..."
+echo ""
+./gradlew build